refactor: rename folders/files and consolidate encoding utilities
- seed/ → mnemonic/ + bip32.ts moved to src root
- derive/ → derivation.ts
- wallet-node/ → wallet/, railgun.ts → railgun-wallet.ts
- hash.ts → encoding.ts with formatToByteLength and bigintToHex
- Update all imports across src/ and test/
