+++ src/bip32.ts
@@ -1,5 +1,5 @@
-import { encodeBytes, sha512HMAC } from '../hash'
-import type { KeyNode } from '../types'
+import { encodeBytes, sha512HMAC } from './encoding'
+import type { KeyNode } from './types'
 
 const CURVE_SEED = encodeBytes('babyjubjub seed') // same calculation as current engine
