+++ src/index.ts
@@ -1,10 +1,10 @@
-export * from './derive'
+export * from './derivation'
 export * from './notes'
 export * from './keys'
-export * from './hash'
+export * from './encoding'
 export type * from './types'
-export { WalletNode } from './wallet-node/wallet-node'
-export type { WalletNodes } from './wallet-node/wallet-node'
-export { RailgunWallet } from './wallet-node/railgun'
-export { Mnemonic } from './seed/bip39'
-export { getMasterKeyFromSeed, childKeyDerivationHardened, getPathSegments } from './seed/bip32'
+export { WalletNode } from './wallet/wallet-node'
+export type { WalletNodes } from './wallet/wallet-node'
+export { RailgunWallet } from './wallet/railgun-wallet'
+export { Mnemonic } from './mnemonic'
+export { getMasterKeyFromSeed, childKeyDerivationHardened, getPathSegments } from './bip32'
