--- src/wallet-node/index.ts
@@ -1,2 +1,0 @@
-export * from './railgun'
-export * from './wallet-node'
