+++ src/wallet/index.ts
@@ -1,0 +1,2 @@
+export * from './railgun-wallet'
+export * from './wallet-node'
